Split github fetch out of main.
